// src/features/invoices/types/invoiceList.types.ts
export type InvoiceStatus = "paid" | "pending" | "draft";

export interface InvoiceSumCardType {
  clientName: string;
  invoiceId: string;
  status: InvoiceStatus;
  onClick: () => void;
}

export interface InvoiceDataType {
  id: string;
  createdAt: string;
  paymentDue: string;
  description: string;
  paymentTerms: number;
  clientName: string;
  clientEmail: string;
  status: InvoiceStatus;
  senderAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  clientAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  items: {
    name: string;
    quantity: number;
    price: number;
    total: number;
  }[];
  total: number;
}